Added functionality to disable add/edit ability on StudentsTableComponent.php if teacher's grades are missing. WIP: add student assignment with asset assignment to ensembles domain. WIP: Migrate from log to ses mailing. WIP: Build link to process verification action when link is clicked. WIP: testing for links to subordinate pages. WIP: Policy for director viewing of student pages to only allow teacher to view student. WIP: Policy for director viewing of student pages to prohibit non-verified email of director to view student entered information.

+++ app/Data/aViewData.php
@@ -22,7 +22,7 @@ abstract class aViewData
         $this->viewPage = $this->getViewPage();
 
         //make commonly used variable quickly available
-        $this->versionId = UserConfig::getValue('versionId') ?? 0;
+        $this->versionId = (int) UserConfig::getValue('versionId') ?? 0;
     }
 
     private function buildMethod($str): void
